Refactor : 예외처리 코드 통합 및 리펙토링
