+++ src/main/java/com/acorn/api/exception/global/GlobalApiExceptionHandler.java
@@ -2,21 +2,48 @@ package com.acorn.api.exception.global;
 
 import com.acorn.api.code.common.ApiHttpErrorCode;
 import com.acorn.api.code.response.ApiErrorResponse;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
+@RestControllerAdvice(annotations = RestController.class)
+public class GlobalApiExceptionHandler {
+
+    @ExceptionHandler(AcontainerException.class)
+    public ResponseEntity<ApiErrorResponse> handleAcontainerException(AcontainerException acontainerException) {
+        log.error(" =========================== AcontainerException: {} =========================== ", acontainerException.getMessage());
+
+        ApiErrorResponse response = ApiErrorResponse.builder()
+                .httpStatus(acontainerException.getHttpStatus())
+                .errorDivisionCode(acontainerException.getErrorDivisionCode())
+                .errorMsg(acontainerException.getErrorMsg())
+                .build();
+
+        return ResponseEntity.status(acontainerException.getHttpStatus()).body(response);
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<ApiErrorResponse> handleIllegalArgumentException(IllegalArgumentException ex) {
+        log.error(" =========================== IllegalArgumentException: {} =========================== ", ex.getMessage());
+
+        ApiErrorResponse response = ApiErrorResponse.builder()
+                .httpStatus(ApiHttpErrorCode.BAD_REQUEST_ERROR.getHttpStatus())
+                .errorDivisionCode(ApiHttpErrorCode.BAD_REQUEST_ERROR.getErrorDivisionCode())
+                .errorMsg(ApiHttpErrorCode.BAD_REQUEST_ERROR.getErrorMsg())
+                .build();
+
+        return ResponseEntity
+                .status(ApiHttpErrorCode.BAD_REQUEST_ERROR.getHttpStatus())
+                .body(response);
+    }
 
     @ExceptionHandler(AuthenticationException.class)
-    public ResponseEntity<ApiErrorResponse> handleAuthenticationException(AuthenticationException ex, HttpServletRequest request) {
-
+    public ResponseEntity<ApiErrorResponse> handleAuthenticationException(AuthenticationException ex) {
         log.error(" =========================== AuthenticationException: {} =========================== ", ex.getMessage());
 
         ApiErrorResponse response = ApiErrorResponse.builder()
@@ -31,7 +58,7 @@ public class GlobalExceptionHandler {
     }
 
     @ExceptionHandler(AccessDeniedException.class)
-    public ResponseEntity<ApiErrorResponse> handleAccessDeniedException(AccessDeniedException ex, HttpServletRequest request) {
+    public ResponseEntity<ApiErrorResponse> handleAccessDeniedException(AccessDeniedException ex) {
 
         log.error(" =========================== AccessDeniedException: {} =========================== ", ex.getMessage());
 
@@ -46,21 +73,8 @@ public class GlobalExceptionHandler {
                 .body(response);
     }
 
-    @ExceptionHandler(AcontainerException.class)
-    public ResponseEntity<ApiErrorResponse> handleAcontainerException(AcontainerException ex, HttpServletRequest request) {
-        log.error(" =========================== AcontainerException: {} =========================== ", ex.getMessage());
-
-        ApiErrorResponse response = ApiErrorResponse.builder()
-                .httpStatus(ex.getHttpStatus())
-                .errorDivisionCode(ex.getErrorDivisionCode())
-                .errorMsg(ex.getErrorMsg())
-                .build();
-
-        return ResponseEntity.status(ex.getHttpStatus()).body(response);
-    }
-
     @ExceptionHandler(Exception.class)
-    public ResponseEntity<ApiErrorResponse> handleUnexpectedException(Exception ex, HttpServletRequest request) {
+    public ResponseEntity<ApiErrorResponse> handleUnexpectedException(Exception ex) {
 
         log.error(" =========================== Unexpected Exception: {} =========================== ", ex.getMessage(), ex);
