+++ src/main/java/com/acorn/api/exception/global/GlobalSqlExceptionHandler.java
@@ -2,63 +2,46 @@ package com.acorn.api.exception.global;
 
 import com.acorn.api.code.common.ApiHttpErrorCode;
 import com.acorn.api.code.response.ApiErrorResponse;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import java.sql.SQLIntegrityConstraintViolationException;
 import java.sql.SQLSyntaxErrorException;
 
 @Slf4j
-@RestControllerAdvice
+@RestControllerAdvice(annotations = RestController.class)
+@RequiredArgsConstructor
 public class GlobalSqlExceptionHandler {
 
     @ExceptionHandler(SQLSyntaxErrorException.class)
-    public ResponseEntity<ApiErrorResponse> handleSQLSyntaxErrorException(SQLSyntaxErrorException ex) {
-
+    public Object handleSQLSyntaxErrorException(SQLSyntaxErrorException ex) {
         log.error(" =========================== SQLSyntaxErrorException: {} =========================== ", ex.getMessage(), ex);
-
-        ApiErrorResponse response = ApiErrorResponse.builder()
-                .httpStatus(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
-                .errorDivisionCode(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getErrorDivisionCode())
-                .errorMsg(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getErrorMsg())
-                .build();
-
-        return ResponseEntity
-                .status(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
-                .body(response);
+        return handleException();
     }
 
     @ExceptionHandler(SQLIntegrityConstraintViolationException.class)
-    public ResponseEntity<ApiErrorResponse> handleSQLIntegrityConstraintViolationException(SQLIntegrityConstraintViolationException ex) {
-
+    public Object handleSQLIntegrityConstraintViolationException(SQLIntegrityConstraintViolationException ex) {
         log.error(" =========================== SQLIntegrityConstraintViolationException: {} =========================== ", ex.getMessage(), ex);
-
-        ApiErrorResponse response = ApiErrorResponse.builder()
-                .httpStatus(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
-                .errorDivisionCode(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getErrorDivisionCode())
-                .errorMsg(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getErrorMsg())
-                .build();
-
-        return ResponseEntity
-                .status(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
-                .body(response);
+        return handleException();
     }
 
     @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<ApiErrorResponse> handleDataIntegrityViolationException(DataIntegrityViolationException ex) {
+    public Object handleDataIntegrityViolationException(DataIntegrityViolationException ex) {
         log.error(" =========================== DataIntegrityViolationException: {} =========================== ", ex.getMessage(), ex);
+        return handleException();
+    }
 
-        ApiErrorResponse response = ApiErrorResponse.builder()
+    private Object handleException() {
+        ApiErrorResponse errorResponse = ApiErrorResponse.builder()
                 .httpStatus(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
                 .errorDivisionCode(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getErrorDivisionCode())
                 .errorMsg(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getErrorMsg())
                 .build();
-
-        return ResponseEntity
-                .status(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
-                .body(response);
+        return ResponseEntity.status(ApiHttpErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus()).body(errorResponse);
     }
 }
